ctforces_backend/requirements.txt
amqp==2.4.2
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==2.2
django-cors-headers==2.5.2
django-extensions==2.1.6
django-guardian==1.5.0
django-nine==0.1.13
django-ratelimit==2.0.0
django-redis==4.10.0
django-rest-framework-tricks==0.2.9
django-stdimage==4.0.1
djangorestframework==3.9.2
drf-extensions==0.4.0
drf-nested-routers==0.91
flower==0.9.3
gevent==1.4.0
graphviz==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonpointer==2.0
kombu==4.5.0
MarkupSafe==1.1.1
parso==0.4.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.0.0
progressbar2==3.39.3
prompt-toolkit==2.0.7
psycopg2==2.8
ptyprocess==0.6.0
pydot==1.4.1
Pygments==2.3.1
pygraphviz==1.5
pyparsing==2.3.1
python-utils==2.3.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
rfc3987==1.3.8
ruamel.yaml==0.15.91
six==1.12.0
sqlparse==0.3.0
strict-rfc3339==0.7
tornado==6.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.2
